code/kuusi/requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.7
django-cors-headers==4.3.1
django-cprofile-middleware==1.0.5
django-debug-toolbar==4.3.0
django-filter==24.3
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
djangorestframework==3.15.2
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
gunicorn==22.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.0
termcolor==2.4.0
tzdata==2024.2
user-agents==2.2.0
requests==2.32.3
langcodes==3.5.0